passing RemoteRepositoryTest

+++ app/src/main/java/com/example/paypay/api/LatestExchangeRateResponse.kt
@@ -2,5 +2,5 @@ package com.example.paypay.api
 
 data class LatestExchangeRateResponse(
     val timestamp: Long,
-    val rates: Map<String, Float>
+    val rates: Map<String, Double>
 )
